Refactor Spark Schemas
