+++ src/sdk/python/rtdip_sdk/pipelines/sources/spark/kinesis.py
@@ -17,7 +17,8 @@ from py4j.protocol import Py4JJavaError
 from pyspark.sql import DataFrame, SparkSession
 from ..interfaces import SourceInterface
 from ..._pipeline_utils.models import Libraries, SystemType
-from ..._pipeline_utils.constants import DEFAULT_PACKAGES, KINESIS_SCHEMA
+from ..._pipeline_utils.constants import DEFAULT_PACKAGES
+from ..._pipeline_utils.spark import KINESIS_SCHEMA
 
 class SparkKinesisSource(SourceInterface):
     '''
